chore: update location runCheck
