+++ src/qa-certification-workspace/qa-certification-checks.service.ts
@@ -47,65 +47,7 @@ export class QACertificationChecksService {
     let errors: string[] = [];
     let locations: LocationIdentifiers[] = [];
 
-    const addLocation = (i: any) => {
-      const systemIDs = [];
-      const componentIDs = [];
-
-      const location = locations.find(
-        l => l?.unitId === i?.unitId || l?.stackPipeId === i?.stackPipeId,
-      );
-
-      if (location) {
-        if (
-          i.monitoringSystemID &&
-          !location.systemIDs.includes(i.monitoringSystemID)
-        ) {
-          location.systemIDs.push(i.monitoringSystemID);
-        }
-        if (i.componentID && !location.componentIDs.includes(i.componentID)) {
-          location.componentIDs.push(i.componentID);
-        }
-      } else {
-        if (i.monitoringSystemID) systemIDs.push(i.monitoringSystemID);
-        if (i.componentID) componentIDs.push(i.componentID);
-
-        locations.push({
-          unitId: i.unitId,
-          locationId: null,
-          stackPipeId: i.stackPipeId,
-          systemIDs,
-          componentIDs,
-        });
-      }
-    };
-
-    if (payload.testSummaryData) {
-      payload.testSummaryData.forEach(i => addLocation(i));
-    }
-
-    if (payload.certificationEventData) {
-      payload.certificationEventData.forEach(i => addLocation(i));
-    }
-
-    if (payload.testExtensionExemptionData) {
-      payload.testExtensionExemptionData.forEach(i => addLocation(i));
-    }
-
-    console.log('locations');
-    console.log(locations);
-
-    // IMPORT-13 Result
-    if (locations.length === 0) {
-      errorList.push(
-        'There are no test summary, certifications events, or extension/exmeption records present in the file to be imported',
-      );
-      this.throwIfErrors(errorList);
-    }
-
-    [locations, errors] = await this.locationChecksService.runChecks(
-      payload.orisCode,
-      locations,
-    );
+    [locations, errors] = await this.locationChecksService.runChecks(payload);
     errorList.push(...errors);
     this.throwIfErrors(errorList);
